feat: :lipstick: header finished
feat: :lipstick: The first section of the page is finished

Images and texts are added and the information is organized in the corresponding area.

feat: :lipstick: The "How to play" area is finished.

Images and texts are added and the information is ordered correctly

feat: :lipstick: Support section is finished

Images and texts are added and the design of the support section is finalized

feat: :lipstick: About us is finished

feat: :lipstick: footer is finished

feat: :sparkles: Work is underway to make it work properly on mobile devices

feat: :sparkles: Added functionality for 430px devices

feat: :sparkles: Mobile functionality is being worked on